DB: replace old DbProvider with DbMovasProvider
